requirements.txt
pysrt
pycorrector
opencc-python-reimplemented
textblob
tqdm
pyyaml
